+++ src/Message/Serializers/AvroEncoder.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Junges\Kafka\Message\Encoders;
+namespace Junges\Kafka\Message\Serializers;
 
 use AvroSchema;
 use FlixTech\AvroSerializer\Objects\RecordSerializer;
