+++ src/Message/Serializers/NullSerializer.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Junges\Kafka\Message\Serializers;
+
+class NullSerializer
+{
+
+}
